fix obs_space

+++ craftax/craftax/envs/craftax_symbolic_env.py
@@ -102,7 +102,7 @@ class CraftaxSymbolicEnvNoAutoReset(EnvironmentNoAutoReset):
             0.0,
             1.0,
             (obs_shape,),
-            dtype=jnp.int32,
+            dtype=jnp.float32,
         )
 
 
@@ -174,5 +174,5 @@ class CraftaxSymbolicEnv(environment.Environment):
             0.0,
             1.0,
             (obs_shape,),
-            dtype=jnp.int32,
+            dtype=jnp.float32,
         )
